move into core/notification.h & .cpp
