+++ src/core/notification.h
@@ -2,9 +2,8 @@
 // Copyright(c) 2023 Intel Corporation. All Rights Reserved.
 #pragma once
 
-#include <librealsense2/h/rs_types.h>
+#include <librealsense2/hpp/rs_types.hpp>
 #include <rsutils/concurrency/concurrency.h>
-#include <rsutils/easylogging/easyloggingpp.h>
 #include <memory>
 #include <string>
 
@@ -14,16 +13,10 @@ namespace librealsense {
 
 struct notification
 {
-    notification( rs2_notification_category category, int type, rs2_log_severity severity, std::string const & description )
-        : category( category )
-        , type( type )
-        , severity( severity )
-        , description( description )
-    {
-        timestamp = std::chrono::duration< double, std::milli >( std::chrono::system_clock::now().time_since_epoch() )
-                        .count();
-        LOG_INFO( description );
-    }
+    notification( rs2_notification_category category,
+                  int type,
+                  rs2_log_severity severity,
+                  std::string const & description );
 
     rs2_notification_category category;
     int type;
