+++ src/core/serialization.h
@@ -11,6 +11,7 @@
 #include "notification.h"
 #include "frame-holder.h"
 #include "stream-profile-interface.h"
+#include "notification.h"
 
 
 namespace librealsense
