+++ src/error-handling.cpp
@@ -1,5 +1,6 @@
 // License: Apache 2.0. See LICENSE file in root directory.
 // Copyright(c) 2019 Intel Corporation. All Rights Reserved.
+
 #include "error-handling.h"
 #include "core/notification.h"
 #include "librealsense-exception.h"
